% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Extract data from import_file() output}
\usage{
get_data(my_file)
}
\arguments{
\item{my_file}{a list produced from import_file()}
}
\value{
tibble
}
\description{
Extract data from import_file() output
}
\details{
This function extracts the data from the list produced by import_file().
The dataframe is standardized and includes six columns.

RID: unique record indetification string

FileName: the name of the file the data were imported from

PlotID: the unique plot identification srting

DateTime: the date and time of the measurement

Element: the measurement type

Value: the measurement.
}
\examples{
library(dataProcessor)
file.list <- list.files(path = "./inst/raw_data", pattern = ".csv", full.names = T, recursive = F)
my_file <- import_file(file.list[1])
get_data(my_file)

}
