% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Extract sampling details from import_file() output}
\usage{
get_details(my_file, my_dat)
}
\arguments{
\item{my_file}{a list produced from import_file()}

\item{my_dat}{a data frame produced from get_data()}
}
\value{
tibble
}
\description{
Extract sampling details from import_file() output
}
\details{
This function produces a data frame of details about the sampling event.
Some of these details are scraped from the Details column of the *.csv produced by HOBOware.
Other's are scraped from the file name, and some are calculated form the data.
}
\examples{
library(dataProcessor)
file.list <- list.files(path = "./inst/raw_data", pattern = ".csv", full.names = T, recursive = F)
my_file <- import_file(file.list[1])
my_data <- get_data(my_file)
get_details(my_file, my_data)

}
