% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_product}
\alias{get_product}
\title{Extract Hobo product and associated information}
\usage{
get_product(my_file)
}
\arguments{
\item{my_file}{}
}
\value{
tibble
}
\description{
Extract Hobo product and associated information
}
