% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Imports *.csv file produced by Onset HOBOware}
\usage{
import_file(
  this_file,
  datestamp_loc = 1,
  plotid_loc = 2,
  plotid_s = 1,
  plotid_e = 3
)
}
\arguments{
\item{this_file}{a character string with the file name, can include directory path.}

\item{datestamp_loc}{a number for the location of the date stamp (an index). Default is 1.}

\item{plotid_loc}{a number for the location of the Plot ID (an index). Default is 2.}

\item{plotid_s}{a number for the beginning of the Plot ID (an index). Default is 1.}

\item{plotid_e}{a number for the end of the Plot ID (an index). Default is 3.}
}
\value{
list
}
\description{
This function produces a list that will be used in get_data() and get_details().
}
\details{
This function finds variables in the file name and the text file to properly
import the *.csv into R.

It returns a list with two components.

The first component is named file_info.
It is a vector that contains the file name, the date stamp on the file name,
the plot ID from the file name, and the number of rows to skip to properly
import the data.

The second component is is a data frame consisting of all the raw data and
metadata.
}
\examples{
file.list <- list.files(path = "./inst/raw_data", pattern = ".csv", full.names = T, recursive = F)
import_file(file.list[1])

}
