% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wxdat.R
\name{import_wxdat}
\alias{import_wxdat}
\title{Wrapper for import_file(), get_data() and get_details()}
\usage{
import_wxdat(this_file, ...)
}
\arguments{
\item{this_file}{a character string with the file name, can include directory path.}

\item{...}{Other arguments to pass to import_wxdat(). These might include the indeces for the
date stamp or plot ID.}
}
\value{
list
}
\description{
This function combines all three HOBOware functions and produces a list
containing the details of the sampling event from get_details() and the
standardized data frame from get_data().
}
\details{
Something, something...
}
\examples{
library(dataProcessor)
file.list <- list.files(path = "./inst/raw_data", pattern = ".csv", full.names = T, recursive = F)
import_wxdat(file.list[1])

}
\seealso{
import_file(), get_data()
}
