% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xls.R
\name{import_xls}
\alias{import_xls}
\title{Import LTVMP monitoring data (Excel workbooks) in R}
\usage{
import_xls(my_xls)
}
\arguments{
\item{my_xls}{A character string of the complete file path of the *.xls
file.}
}
\value{
A list with three (3) objects.
\describe{
\item{\strong{file_info}}{ This component is a vector that contains the
file name, today's date, plot ID, and year the plot was sampled.}
\item{\strong{sampling_event}}{ This component is a vector that contains
the year the plot was sampled, the plot ID, the sample date, and the
names of the observers.}
\item{\strong{data}}{ This component is a data frame in long format.}
}
}
\description{
This function imports SEUG long-term vegetation monitoring data from
Microsoft Excel workbooks into R.
}
\details{
This function pulls information and data from the "SiteInfo" and
"Data" sheets of the SEUG LTVMP Excel workbook and mungesthem from wide
to long format.
}
\seealso{
\code{\link[readxl]{read_excel}}
}
