% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hobo_2008.R
\name{export_hobo_2008}
\alias{export_hobo_2008}
\title{Export Onset Hobo data to Access Database}
\usage{
export_hobo_2008(
  my_file,
  my_db,
  import_table,
  raw_data_table,
  prcp_data_table,
  temp_rh_data_table,
  details_table,
  verbose = TRUE,
  view = TRUE
)
}
\arguments{
\item{my_file}{A character string of the complete file path of your *.csv
file.}

\item{my_db}{A connected database from \code{\link{RODBC}}.}

\item{import_table}{A character string of the name of the import log table.}

\item{raw_data_table}{A character string of the name of the raw data table.}

\item{prcp_data_table}{A character string of the name of the processed
precipitation data table.}

\item{temp_rh_data_table}{A character string of the name of the processed
temperature and relative humidity data table.}

\item{details_table}{A character string of the name of the logger details
table.}

\item{verbose}{Logical. Show messages showing progress. Default is TRUE. If
FALSE, messages are suppressed.}

\item{view}{Logical. Prints data to console before writing them to the
database. Default is TRUE. If FALSE, data are not printed and there is no
prompt before writing data to the database.}
}
\value{
Data is written to database tables. Objects are not returned.
}
\description{
This function uses the \href{https://github.com/scoyoc/raindancer}{raindancer}
package to processes data from Onset HOBO loggers used in the SEUG LTMVP
from 2008-2019. It then exports the data to a Microsoft Access database.
}
\details{
This function uses two functions from the raindacer package.
\code{\link[raindancer]{import_hobo_2008}} is usec to read Hobo data in
into R, and then \code{\link[raindancer]{process_hobo_2008}} is used to
summarize the data. The processed data are then exported to a connected
Microsoft Access database.
}
\examples{
\dontrun{
library("raindancer")
library("dataprocessR")

# Connect to DB
my_db <- RODBC::odbcConnectAccess2007("C:/path/to/database.accdb")

# List files
my_dir <- "C:/path/to/data"
file_list <- list.files(my_dir, pattern = ".csv", full.names = TRUE,
                        recursive = FALSE)
# Select file
my_file <- file_list[10]

# Process file and save to database
export_hobo_2008(my_file = my_file, my_db = my_db,
                 import_table = "tblWxImportLog",
                 raw_data_table = "tblWxData_raw",
                 prcp_data_table = "tblWxData_PRCP",
                 temp_rh_data_table = "tblWxData_TEMP_RH",
                 details_table = "tblWxLoggerDetails")
}
}
\seealso{
\href{https://github.com/scoyoc/raindancer}{raindancer},
\code{\link[raindancer]{import_hobo_2008}},
\code{\link[raindancer]{raindance}}, \code{\link[raindancer]{sundance}},
\code{\link[raindancer]{process_hobo}}, \code{\link{RODBC}},
\code{\link[RODBC]{sqlSave}}, \code{\link[RODBC]{odbcConnectAccess2007}}
}
